# docs/swagger.yaml
basePath: /
definitions:
  handler.ResponseSuccess:
    properties:
      data: {}
      message:
        type: string
    type: object
  handler.errorResponse:
    properties:
      message:
        type: string
    type: object
  handler.errorResponseData:
    properties:
      data: {}
      message:
        type: string
    type: object
...
  model.NewsFull:
    properties:
      id:
        type: integer
      post_body:
        type: string
      post_body_ru:
        type: string
      post_date:
        type: string
      post_img_path:
        type: string
      post_img_url:
        type: string
      post_title:
        type: string
      post_title_ru:
        type: string
    type: object
  model.NewsPost:
    properties:
      post_body:
        type: string
      post_body_ru:
        type: string
      post_img_url:
        type: string
      post_title:
        type: string
      post_title_ru:
        type: string
    type: object
...
  model.allNews:
    properties:
      allHome:
        items:
          $ref: '#/definitions/model.NewsFull'
        type: array
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API Server for Norbekov App
  title: Norbekov API
  version: "1.0"
paths:
...
  /api/news/create:
    post:
      consumes:
      - application/json
      description: create news_post
      operationId: create-news_post
      parameters:
      - description: home info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.NewsPost'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.errorResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create NewsPost
      tags:
      - News
  /api/news/delete:
    delete:
      consumes:
      - application/json
      description: delete news
      operationId: delete-news
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.NewsFull'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Delete news
      tags:
      - News
  /api/news/update/{id}:
    put:
      consumes:
      - application/json
      description: Update news by id
      operationId: update-news-id
      parameters:
      - description: Param ID
        in: path
        name: id
        required: true
        type: integer
      - description: home info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.NewsPost'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.errorResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update  News By ID
      tags:
      - News
  /api/news/upload-img/{id}:
    patch:
      consumes:
      - application/json
      - multipart/form-data
      description: Upload News Image
      operationId: uploadImg
      parameters:
      - description: Param ID
        in: path
        name: id
        required: true
        type: integer
      - description: file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      - application/octet-stream
      - image/png
      - image/jpeg
      - image/jpg
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Upload News Image
      tags:
      - News
...
  /news/get:
    get:
      consumes:
      - application/json
      description: get news post
      operationId: get-news
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.NewsFull'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Get news
      tags:
      - News
  /newsS/get:
    get:
      consumes:
      - application/json
      description: get  news
      operationId: get-newsS
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.allNews'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Get newS
      tags:
      - News
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"